+++ lib/src/domain/observatory_thread.dart
@@ -1,5 +1,0 @@
-enum ObservatoryThread {
-  unspecified,
-  foreground,
-  background,
-}