+++ lib/src/feature/_common/domain/ports/observation_history.dart
@@ -1,0 +1,6 @@
+import 'package:observatory/src/feature/_common/domain/log_level.dart';
+import 'package:observatory/src/feature/_common/domain/observation.dart';
+
+abstract interface class ObservationHistory {
+  List<Observation> last({required int limit, LogLevel minimumLevel = LogLevel.verbose});
+}